Implement a "retry" for net.Dial calls.
After some researches in roflcoptor issue, it appears procsnitchd isn't
starting fast enough, and doesn't provide the required socket for
procsnitchd client to work properly.

Implementing a "retry" in procsnitchd client lib should allow
a better handling for that kind of situation.

Small note: I'm not fluent at all with Go, and this PR might even
not work (don't know how to test).
It's just an example, a possible way to follow.